requirements.txt
# This requirements file is a commented mirror of bloodmallet/requirements.txt, which is needed where it is for deployment.
# Furthermore due to the deployment the requirements file can't have comments for now.
# (It would break the deployment)

# Attention!
# Make sure to use 'python -m pip install --upgrade pip setuptools wheel' before trying to use 'pip install -U -r requirements.txt'

# Django~=2.1.2
Django~=3.0
django-allauth
requests

# libsass is not required on the production environment, if SASS/SCSS files have been precompiled and deployed using offline compilation.
libsass
# django-compressor is required only for offline compilation, when using the command `manage.py compilescss`.
django-compressor
django-sass-processor

# autogenerate forms
django-crispy-forms

# google api for the secret app
google-api-python-client

# google example uses this adapter for mysql
pymysql

# handles storage of files from models in google storage bucket
django-storages[google]

# add DB models to translations
django-vinaigrette

# hopefully enables us to deliver json data to foreigners
django-cors-headers
